moves the keyboard macro implementation
