--- rmk/src/config/keyboard_macros/mod.rs
@@ -1,2 +1,0 @@
-pub mod keyboard_macro;
-pub mod macro_config;
