+++ rmk/src/keyboard_macros.rs
@@ -6,7 +6,7 @@ use crate::{
     via::keycode_convert::{from_ascii, to_ascii},
 };
 
-use crate::config::keyboard_macros::macro_config::MACRO_SPACE_SIZE;
+use crate::config::macro_config::MACRO_SPACE_SIZE;
 
 /// encoded with the two bytes, content at the third byte
 /// 0b 0000 0001 1000-1010 (VIAL_MACRO_EXT) are not supported
