+++ rmk/src/storage/mod.rs
@@ -24,7 +24,7 @@ use self::eeconfig::EeKeymapConfig;
 use crate::action::{EncoderAction, KeyAction};
 use crate::channel::FLASH_CHANNEL;
 use crate::combo::{Combo, COMBO_MAX_LENGTH, COMBO_MAX_NUM};
-use crate::config::keyboard_macros::macro_config::MACRO_SPACE_SIZE;
+use crate::config::macro_config::MACRO_SPACE_SIZE;
 use crate::config::StorageConfig;
 use crate::fork::{Fork, StateBits, FORK_MAX_NUM};
 use crate::hid_state::{HidModifiers, HidMouseButtons};
