+++ rmk/src/via/mod.rs
@@ -9,7 +9,7 @@ use num_enum::{FromPrimitive as _, TryFromPrimitive as _};
 use protocol::{ViaCommand, ViaKeyboardInfo, VIA_FIRMWARE_VERSION, VIA_PROTOCOL_VERSION};
 use vial::process_vial;
 
-use crate::config::keyboard_macros::macro_config::MACRO_SPACE_SIZE;
+use crate::config::macro_config::MACRO_SPACE_SIZE;
 use crate::config::VialConfig;
 use crate::hid::{HidError, HidReaderTrait, HidWriterTrait};
 use crate::keymap::KeyMap;
